weekly_update/requirements.txt
attrs==23.2.0
black==24.3.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
idna==3.6
isodate==0.6.1
lxml==5.2.1
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.2.0
mypy-extensions==1.0.0
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pathspec==0.12.1
pendulum==3.0.0
pip-system-certs==4.0
platformdirs==4.2.0
pycparser==2.22
Pygments==2.17.2
PyJWT==2.8.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
requests-file==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
shellingham==1.5.4
simple-salesforce==1.12.6
six==1.16.0
time-machine==2.14.1
typer==0.12.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wrapt==1.16.0
zeep==4.2.1
